Improved the update user functionality and updated the tests
