+++ api/utils/data_validation.js
@@ -11,3 +11,18 @@ exports.isEmailAlreadyInUse = async (email) => {
 
     return user;
 };
+
+exports.formatUserData = (user) => {
+    const signedInUser = {
+        email: user.email,
+        username: user.username,
+        chestCircumference: user.chestCircumference ?? undefined,
+        armLength: user.armLength ?? undefined,
+        armCircumference: user.armCircumference ?? undefined,
+        bodyLength: user.bodyLength ?? undefined,
+        shoulderWidth: user.shoulderWidth ?? undefined,
+        preferredUnit: user.preferredUnit ?? undefined,
+    };
+    
+    return signedInUser;
+};
